F/23 snippets

+++ ProjectManager/Pages/Base/PagesManager.xaml.cs
@@ -2,6 +2,8 @@
 using ProjectManager.Data.Dashboard;
 using ProjectManager.Pages.Events;
 using ProjectManager.Pages.Settings;
+using ProjectManager.Pages.Snippets;
+using SnippetsManager.Models;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -28,7 +30,8 @@ namespace ProjectManager.Pages.Base
         private static readonly List<Type> RESTRICTED_RECENT_PAGES = new List<Type>()
         {
             typeof(DashboardPage),
-            typeof(SettingsPage)
+            typeof(SettingsPage),
+            typeof(SnippetsEditPage),
         };
 
         private static readonly Dictionary<string, Type> PAGES_NAMES_MAPPING = new Dictionary<string, Type>()
@@ -38,6 +41,9 @@ namespace ProjectManager.Pages.Base
             { nameof(SettingsDatabasesPage), typeof(SettingsDatabasesPage) },
             { nameof(SettingsInfoPage), typeof(SettingsInfoPage) },
             { nameof(SettingsPage), typeof(SettingsPage) },
+            { nameof(SettingsSnippetsPage), typeof(SettingsSnippetsPage) },
+            { nameof(SnippetsEditPage), typeof(SnippetsEditPage) },
+            { nameof(SnippetsPage), typeof(SnippetsPage) },
         };
 
 
@@ -140,11 +146,16 @@ namespace ProjectManager.Pages.Base
         {
             if (CanGoBack)
             {
-                var currPageIndex = _loadedPages.IndexOf(LoadedPage);
+                var currPage = LoadedPage;
+
+                var currPageIndex = _loadedPages.IndexOf(currPage);
                 var destPageIndex = Math.Max(0, currPageIndex - backCount);
 
                 //  Get previous page from list to load into content frame.
                 var destPage = _loadedPages[destPageIndex];
+
+                if (currPage != null && !currPage.OnGoBackFromPage(destPage))
+                    return;
 
                 //  Remove other pages loaded further.
                 _loadedPages.RemoveRange(destPageIndex + 1, PagesCount - (destPageIndex + 1));
@@ -153,7 +164,7 @@ namespace ProjectManager.Pages.Base
                 contentFrame.Navigate(destPage);
 
                 //  Invoke external event.
-                var args = new PageLoadedEventArgs(destPage);
+                var args = new PageLoadedEventArgs(currPage, destPage);
                 OnPageBack?.Invoke(this, args);
 
                 OnPropertyChanged(nameof(CanGoBack));
@@ -170,6 +181,11 @@ namespace ProjectManager.Pages.Base
 
             if (pageToLoad != null)
             {
+                var currPage = LoadedPage;
+
+                if (currPage != null && !currPage.OnGoForwardFromPage(page))
+                    return;
+
                 //  Add page to history.
                 _loadedPages.Add(page);
 
@@ -177,7 +193,7 @@ namespace ProjectManager.Pages.Base
                 contentFrame.Navigate(pageToLoad);
 
                 //  Invoke external event.
-                var args = new PageLoadedEventArgs(page);
+                var args = new PageLoadedEventArgs(currPage, page);
                 OnPageLoaded?.Invoke(this, args);
 
                 //  Add page to recently used items.
@@ -231,7 +247,7 @@ namespace ProjectManager.Pages.Base
         #region STATIC PAGES LOADERS METHODS
 
         //  --------------------------------------------------------------------------------
-        /// <summary> Load page by it's name. </summary>
+        /// <summary> Load and show page by it's name. </summary>
         /// <param name="pageName"> Page name. </param>
         /// <returns> Loaded page. </returns>
         public BasePage LoadPageByName(string pageName)
@@ -250,7 +266,7 @@ namespace ProjectManager.Pages.Base
         }
 
         //  --------------------------------------------------------------------------------
-        /// <summary> Load Settings Page (Appearance). </summary>
+        /// <summary> Load and show Settings Page (Appearance). </summary>
         /// <returns> Loaded page. </returns>
         public BasePage LoadSettingsAppearancePage()
         {
@@ -260,7 +276,7 @@ namespace ProjectManager.Pages.Base
         }
 
         //  --------------------------------------------------------------------------------
-        /// <summary> Load Settings Page (Database). </summary>
+        /// <summary> Load and show Settings Page (Database). </summary>
         /// <returns> Loaded page. </returns>
         public BasePage LoadSettingsDatabasePage()
         {
@@ -270,7 +286,7 @@ namespace ProjectManager.Pages.Base
         }
 
         //  --------------------------------------------------------------------------------
-        /// <summary> Load Settings Page (Informations). </summary>
+        /// <summary> Load and show Settings Page (Informations). </summary>
         /// <returns> Loaded page. </returns>
         public BasePage LoadSettingsInformationsPage()
         {
@@ -280,11 +296,21 @@ namespace ProjectManager.Pages.Base
         }
 
         //  --------------------------------------------------------------------------------
-        /// <summary> Load SettingsPage (General). </summary>
+        /// <summary> Load and show SettingsPage (General). </summary>
         /// <returns> Loaded page. </returns>
         public BasePage LoadSettingsMainPage()
         {
             var page = new SettingsPage(this);
+            LoadPage(page);
+            return page;
+        }
+
+        //  --------------------------------------------------------------------------------
+        /// <summary> Load and show Settings Page (Snippets). </summary>
+        /// <returns> Loaded page. </returns>
+        public BasePage LoadSettingsSnippetsPage()
+        {
+            var page = new SettingsSnippetsPage(this);
             LoadPage(page);
             return page;
         }
@@ -295,6 +321,27 @@ namespace ProjectManager.Pages.Base
         public BasePage LoadDashboardPage()
         {
             var page = new DashboardPage(this);
+            LoadPage(page);
+            return page;
+        }
+
+        //  --------------------------------------------------------------------------------
+        /// <summary> Load and show SnippetsPage. </summary>
+        /// <returns> Loaded page. </returns>
+        public BasePage LoadSnippetsPage()
+        {
+            var page = new SnippetsPage(this);
+            LoadPage(page);
+            return page;
+        }
+
+        //  --------------------------------------------------------------------------------
+        /// <summary> Load and show SnippetsEditPage. </summary>
+        /// <param name="snippetItem"> Snippet item to edit. </param>
+        /// <returns> Loaded page. </returns>
+        public BasePage LoadSnippetsEditPage(SnippetItem snippetItem)
+        {
+            var page = new SnippetsEditPage(this, snippetItem);
             LoadPage(page);
             return page;
         }
